rutinas/scripts/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.0
supabase==2.3.0
python-dotenv==1.0.1
lxml==5.1.0
openpyxl==3.1.2
scikit-learn==1.4.1.post1
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
